+++ packages/common/src/index.ts
@@ -1,5 +1,6 @@
 export * from './url';
 export * from './not';
+export * from './mix';
 export * from './rip';
 export * from './unit';
 export * from './hash';